Module Delegate_events.Baking_forge

include Tezos_base__TzPervasives.Internal_event.Simple

type 'a t

Event declarations where 'a is the type of the event parameters.

val emit : 'a t -> 'a -> unit Lwt.t

Emit an instance of an event.

val emit__dont_wait__use_with_care : 'a t -> 'a -> unit

Emit an instance of an event but do not wait for completion. Only use if you really need to not wait for logging resolution before continuing. May cause increased memory usage and out-of-order log messages.

val declare_0 : ?⁠section:string list -> name:string -> msg:string -> ?⁠level:Tezos_event_logging.Internal_event.level -> unit -> unit t

Declare an event with no parameters.

val declare_1 : ?⁠section:string list -> name:string -> msg:string -> ?⁠level:Tezos_event_logging.Internal_event.level -> ?⁠pp1:(Stdlib.Format.formatter -> 'a -> unit) -> (string * 'a Data_encoding.t) -> 'a t

Declare an event with one parameter.

val declare_2 : ?⁠section:string list -> name:string -> msg:string -> ?⁠level:Tezos_event_logging.Internal_event.level -> ?⁠pp1:(Stdlib.Format.formatter -> 'a -> unit) -> (string * 'a Data_encoding.t) -> ?⁠pp2:(Stdlib.Format.formatter -> 'b -> unit) -> (string * 'b Data_encoding.t) -> ('a * 'b) t

Declare an event with two parameters.

val declare_3 : ?⁠section:string list -> name:string -> msg:string -> ?⁠level:Tezos_event_logging.Internal_event.level -> ?⁠pp1:(Stdlib.Format.formatter -> 'a -> unit) -> (string * 'a Data_encoding.t) -> ?⁠pp2:(Stdlib.Format.formatter -> 'b -> unit) -> (string * 'b Data_encoding.t) -> ?⁠pp3:(Stdlib.Format.formatter -> 'c -> unit) -> (string * 'c Data_encoding.t) -> ('a * 'b * 'c) t

Declare an event with three parameters.

val declare_4 : ?⁠section:string list -> name:string -> msg:string -> ?⁠level:Tezos_event_logging.Internal_event.level -> ?⁠pp1:(Stdlib.Format.formatter -> 'a -> unit) -> (string * 'a Data_encoding.t) -> ?⁠pp2:(Stdlib.Format.formatter -> 'b -> unit) -> (string * 'b Data_encoding.t) -> ?⁠pp3:(Stdlib.Format.formatter -> 'c -> unit) -> (string * 'c Data_encoding.t) -> ?⁠pp4:(Stdlib.Format.formatter -> 'd -> unit) -> (string * 'd Data_encoding.t) -> ('a * 'b * 'c * 'd) t

Declare an event with four parameters.

val declare_5 : ?⁠section:string list -> name:string -> msg:string -> ?⁠level:Tezos_event_logging.Internal_event.level -> ?⁠pp1:(Stdlib.Format.formatter -> 'a -> unit) -> (string * 'a Data_encoding.t) -> ?⁠pp2:(Stdlib.Format.formatter -> 'b -> unit) -> (string * 'b Data_encoding.t) -> ?⁠pp3:(Stdlib.Format.formatter -> 'c -> unit) -> (string * 'c Data_encoding.t) -> ?⁠pp4:(Stdlib.Format.formatter -> 'd -> unit) -> (string * 'd Data_encoding.t) -> ?⁠pp5:(Stdlib.Format.formatter -> 'e -> unit) -> (string * 'e Data_encoding.t) -> ('a * 'b * 'c * 'd * 'e) t

Declare an event with five parameters.

val declare_6 : ?⁠section:string list -> name:string -> msg:string -> ?⁠level:Tezos_event_logging.Internal_event.level -> ?⁠pp1:(Stdlib.Format.formatter -> 'a -> unit) -> (string * 'a Data_encoding.t) -> ?⁠pp2:(Stdlib.Format.formatter -> 'b -> unit) -> (string * 'b Data_encoding.t) -> ?⁠pp3:(Stdlib.Format.formatter -> 'c -> unit) -> (string * 'c Data_encoding.t) -> ?⁠pp4:(Stdlib.Format.formatter -> 'd -> unit) -> (string * 'd Data_encoding.t) -> ?⁠pp5:(Stdlib.Format.formatter -> 'e -> unit) -> (string * 'e Data_encoding.t) -> ?⁠pp6:(Stdlib.Format.formatter -> 'f -> unit) -> (string * 'f Data_encoding.t) -> ('a * 'b * 'c * 'd * 'e * 'f) t

Declare an event with six parameters.

val declare_7 : ?⁠section:string list -> name:string -> msg:string -> ?⁠level:Tezos_event_logging.Internal_event.level -> ?⁠pp1:(Stdlib.Format.formatter -> 'a -> unit) -> (string * 'a Data_encoding.t) -> ?⁠pp2:(Stdlib.Format.formatter -> 'b -> unit) -> (string * 'b Data_encoding.t) -> ?⁠pp3:(Stdlib.Format.formatter -> 'c -> unit) -> (string * 'c Data_encoding.t) -> ?⁠pp4:(Stdlib.Format.formatter -> 'd -> unit) -> (string * 'd Data_encoding.t) -> ?⁠pp5:(Stdlib.Format.formatter -> 'e -> unit) -> (string * 'e Data_encoding.t) -> ?⁠pp6:(Stdlib.Format.formatter -> 'f -> unit) -> (string * 'f Data_encoding.t) -> ?⁠pp7:(Stdlib.Format.formatter -> 'g -> unit) -> (string * 'g Data_encoding.t) -> ('a * 'b * 'c * 'd * 'e * 'f * 'g) t

Declare an event with seven parameters.

val declare_8 : ?⁠section:string list -> name:string -> msg:string -> ?⁠level:Tezos_event_logging.Internal_event.level -> ?⁠pp1:(Stdlib.Format.formatter -> 'a -> unit) -> (string * 'a Data_encoding.t) -> ?⁠pp2:(Stdlib.Format.formatter -> 'b -> unit) -> (string * 'b Data_encoding.t) -> ?⁠pp3:(Stdlib.Format.formatter -> 'c -> unit) -> (string * 'c Data_encoding.t) -> ?⁠pp4:(Stdlib.Format.formatter -> 'd -> unit) -> (string * 'd Data_encoding.t) -> ?⁠pp5:(Stdlib.Format.formatter -> 'e -> unit) -> (string * 'e Data_encoding.t) -> ?⁠pp6:(Stdlib.Format.formatter -> 'f -> unit) -> (string * 'f Data_encoding.t) -> ?⁠pp7:(Stdlib.Format.formatter -> 'g -> unit) -> (string * 'g Data_encoding.t) -> ?⁠pp8:(Stdlib.Format.formatter -> 'h -> unit) -> (string * 'h Data_encoding.t) -> ('a * 'b * 'c * 'd * 'e * 'f * 'g * 'h) t

Declare an event with eight parameters.

val section : string list

val double_bake_near_miss : Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Raw_level.t t

val inject_baked_block : (Tezos_crypto.Block_hash.t * Stdlib.Bytes.t * Tezos_base__TzPervasives.Operation.t list list) t

val baking_local_validation_start : Tezos_crypto.Block_hash.t t

val context_fetch_error : Tezos_base__TzPervasives.error Tezos_base__TzPervasives.TzTrace.trace t

val reopen_context : unit t

val baking_rejected_invalid_operation : (Tezos_crypto.Operation_hash.t * Tezos_base__TzPervasives.error Tezos_base__TzPervasives.TzTrace.trace) t

val shell_prevalidation_notice : unit t

val shell_prevalidation_new_protocol : unit t

val found_valid_operations : (int * int * Tezos_base__TzPervasives.Time.System.t * Tezos_base__TzPervasives.Fitness.t) t

val block_conversion_failed : Tezos_base__TzPervasives.error Tezos_base__TzPervasives.TzTrace.trace t

val block_injection_failed : (Tezos_base__TzPervasives.Operation.t list * Tezos_base__TzPervasives.error Tezos_base__TzPervasives.TzTrace.trace) t

val built_invalid_block_error : Tezos_base__TzPervasives.error Tezos_base__TzPervasives.TzTrace.trace t

val try_baking : (Tezos_crypto.Block_hash.t * int * string * Tezos_base__TzPervasives.Time.System.t) t

val new_head_received : unit t

val client_side_validation_error : Tezos_base__TzPervasives.error Tezos_base__TzPervasives.TzTrace.trace t

val waiting_before_injection : (Tezos_base__TzPervasives.Time.System.t * Tezos_base__TzPervasives.Time.System.t) t

val try_forging : (Tezos_crypto.Block_hash.t * int * string * Tezos_base__TzPervasives.Time.System.t) t

val start_injecting_block : (int * Tezos_base__TzPervasives.Fitness.t * string * Tezos_crypto.Block_hash.t * Tezos_client_base.Client_keys.Public_key_hash.t) t

val injected_block : (Tezos_crypto.Block_hash.t * string * Tezos_crypto.Block_hash.t * Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Raw_level.t * int * Tezos_base__TzPervasives.Fitness.t * Tezos_base__TzPervasives.Operation.t list) t

val baking_slot_fetch_errors : Tezos_base__TzPervasives.error Tezos_base__TzPervasives.TzTrace.trace t

val no_slot_found : (Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Raw_level.t * int) t

val have_baking_slot : (Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Raw_level.t * int * Tezos_base__TzPervasives.Time.System.t * string * Tezos_crypto.Block_hash.t * Tezos_client_base.Client_keys.Public_key_hash.t) t

val read_nonce_fail : Tezos_base__TzPervasives.error Tezos_base__TzPervasives.TzTrace.trace t

val nonce_retrieval_fail : Tezos_base__TzPervasives.error Tezos_base__TzPervasives.TzTrace.trace t

val nonce_injection_fail : Tezos_base__TzPervasives.error Tezos_base__TzPervasives.TzTrace.trace t

val retrying_on_error : Tezos_base__TzPervasives.error Tezos_base__TzPervasives.TzTrace.trace t

val endorsement_received : (int * int) t

val expected_validity_time : (Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Timestamp.t * int) t

val reading_per_block : string t

val per_block_vote_file_notice : string t

val reading_liquidity_baking : unit t

val liquidity_baking_escape_vote : bool t