Module Tezos_baking_010_PtGRANAD__Client_baking_pow

val empty_proof_of_work_nonce : Stdlib.Bytes.t

A null proof-of-work nonce. This should only be used to nonsensical blocks of the correct size and shape.

val mine : Tezos_client_010_PtGRANAD.Protocol_client_context.#full -> Tezos_shell_services.Shell_services.chain -> Tezos_shell_services.Block_services.block -> Tezos_base__TzPervasives.Block_header.shell_header -> (Stdlib.Bytes.t -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Block_header.contents) -> Tezos_raw_protocol_010_PtGRANAD.Alpha_context.Block_header.contents Tezos_base__TzPervasives.tzresult Lwt.t

mine cctxt chain block header builder returns a block with a valid proof-of-work nonce. The function builder, provided by the caller, is used to make the block. All the internal logic of generating nonces and checking for the proof-of-work threshold is handled by mine.